Move files to src/pypwsqc
